setting up model, renamed appropriate files
