--- .env.example
@@ -1,2 +1,0 @@
-MONGODB_URI=mongodb://localhost:27017/
-PORT=3000
