+++ lib/app.js
@@ -7,7 +7,7 @@ app.use(express.static(publicDir));
 app.use(express.json());
 
 //don't forget to change thing:
-const thing = require('./routes/thing');
-app.use('/api/things', thing);
+// const movie = require('./routes/movies');
+// app.use('/api/movies', movie);
 
 module.exports = app;
