+++ lib/models/movie.js
@@ -5,4 +5,4 @@ const schema = new Schema({
     //your schema model
 });
 
-module.exports = mongoose.model('thing', schema);
+module.exports = mongoose.model('Movie', schema);
