+++ server.js
@@ -1,6 +1,6 @@
 require('dotenv').config();
 const { createServer } = require('http');
-require('./lib/mongodb');
+require('./lib/connect');
 const app = require('./lib/app');
 
 const PORT = process.env.PORT || 3000;
