+++ test/e2e/.env
@@ -1,0 +1,1 @@
+MONGODB_URI=mongodb://localhost:27017/testing
