--- test/e2e/.env.example
@@ -1,1 +1,0 @@
-MONGODB_URI=mongodb://localhost:27017/
