--- test/unit/things.test.js
@@ -1,19 +1,0 @@
-const chai = require('chai');
-const { assert } = chai;
-const YourModel = require('../../lib/models/your-model');
-
-describe('YourModel model', () => {
-
-    it('validates good model', () => {
-        const data = {
-            // example full, good data
-        };
-        const yourModel = new YourModel(data);
-
-        const json = yourModel.toJSON();
-        delete json._id;
-        assert.deepEqual(json, data);
-        assert.isUndefined(yourModel.validateSync());
-    });
-
-});
